code reorganized
