+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod awareness;
+pub mod broadcast;
 pub mod sync;
 pub mod ws;
